+++ src/features/mypage/components/admin-container.tsx
@@ -1,0 +1,3 @@
+export const AdminUserContainer = () => {
+  return <>공급자 마이페이지</>;
+};
