+++ src/features/mypage/components/user-container.tsx
@@ -1,0 +1,11 @@
+// 내가 참여한 설문(기록)
+// 설문 바로가기 (링크 복사?)
+// 당첨 여부
+
+export const UserContainer = () => {
+  return (
+    <div>
+      <h1>소비자 마이페이지</h1>
+    </div>
+  );
+};
